Update by local authority view
Previously this view was very slow - taking ~10 seconds to load ~150
projects, the reason is it was making single calls to the Academies API
to fetch the establishment in order to work out which local authority a
project belongs to.

We've since updated the EstablishmentFetcher to handle any number of
projects and so we can use it here to bulk fetch the establishments,
which should speed things up considereably.

This meant some changes in the controller, but nothing that doesn't
leave it better.