Refactor IncomingTrustsFetcher
This is not really an 'incoming trusts fetcher' - we use it to do that,
but it simply fetches trust - so let's call it that and start a naming
convertion as well, ending 'service objects' with 'service'.

We also initialize the object with projects to prevent the case where
two different sets of projects might get passed to the same instance.

Lastly rename 'call' to 'call!' to hint at the side effects that this
produces - which it the goal, a bunch of projects with the
incoming_trust already set.