Add a basic list of users
We'll be managing users in the application, to start things off here we
lay the ground work.

Display a list of the users in the application.

We think Service Support user will be responsible for user management so
we add the navigation there and restrict access to users with that
role.