Update User model
We will soon be letting certain users manage other users within the
application UI. We have also recently introduce the concept of users
belonging to 'teams' with a link between the team you are in and the
roles you are assigned.

This work takes the existing User model and adds behaviour so that only
a users team or team_leader attribute need to be set and we figure out
the roles based on those.

The logic is as follows:

- team: regional_casework_services team_leader: false = caseworker: true
- team: regional_casework_services team_leader: true = caseworker:
  false team_leader: true
- team: one of the regions e.g. london team_leader: false =
  regional_delivery_officer: true
- team: one of the regions e.g. london team_leader: true =
  regional_delivery_officer: true team_leader: true
- team: service support = service_support: true, team_leader: Always
  false

All other teams do not apply any roles right now.

We should rename the roles to be more descriptive, but that is not part
of this work.

We have added various validations that makes sense when users are
allowed to add users view a form.

These validations include a loose @eduction.gov.uk regex so that only
appropriate email addresses can be added. We kept this loose as we have
the second validation and we don't know what education.gov.uk addresses
threre are.