Users can be updated
Allows users to be updated.

We do not want Users to be deleted from the application to retain the
who worked on what.

Note, if a User leaves the DfE they will no longer be in the active
directory and will not be able to sign in to the applciation.

We may want to disable users at some point to cover users moving teams
within the DfE , but that is not part of this work.