Update session controller
When a new session is created we save the Activce Directory IDs for a
user. Now that the User model requires `team` we need to save these
values without validation, otherwise, whilst we have users with no team,
they will not be able to sign in.

Here we swtich from `update!` to `update_attribute`, which is a good
way to update a single value without validaiton:

https://api.rubyonrails.org/classes/ActiveRecord/Persistence.html#method-i-update_attribute