Opening projects pre fetch their data
The projects shown on the opening view are loaded and sorted by a
service object. We use the same service to fetch the projects in the csv
download.

Both the view and csv loading perfromance can be imporved by pre
fetching the project establishments and trusts.

Here we add the prefetching to the service objects. Testing the sorting
with prefetching would requirely a huge amount of stubbing, instead we
opted to have the option to switch off prefetching allowing the sorting
to be tested without prefecthing.

The prefetching is enabled by default and tested in isolation.

Because the service is now prefetching, we had to update a few specs to
account for that.

We also took the oportunity to add `Service` to the name.