Add `ByTeamProjectFetcherService` service
This service fetches projects by their team & the user's team.

If the user's team is `regional_casework_services`, then they see all projects
there the Project.team is `regional_casework_services`

If the user's team is a region, then they see all projects where Project.region
is the same as the user's team (note that it is the project's REGION).

Users in other teams do not see any projects.

The fetcher also filters projects by in-progress, completed and unassigned. It
was easier to do this filtering in the service, where we can use chained scopes
to get the right project set.