Add DfE Frontend styles
The DfE would like us to use their design system:

https://design.education.gov.uk/design-system

Unfortunatley the npm package is out of date and we could not use it
directly, instead we've uncluded the styles from the distribution.

We have added the css to the prettier ignore as we want to use the file
as is without ever modifying it - this is so as it evolves we can drop
in a replacement with little effort.

The frontend has a default width of 1200px, so we have adopted that.

We are not convinced about the choices made in the font stack, but this
is what has been requested.