Use ByTeamProjectFetcherService in the team/projects_controller
Use the new ByTeamProjectFetcherService in the team/projects_controller. This
makes sure the user only sees projects relevant to their team.

We can remove a lot of the team projects feature spec now, because the new
Service spec covers a lot of the logic and does it faster. We only use the
feature spec to make sure we see helpful messages if there are no projects.

We also alter the project assignment spec to reflect that caseworker and
regional delivery officer users cannot see the Unassigned team projects page.

We can no longer use the Pre-fetcher services in the team/projects_controller
because the result of the ByTeamProjectFetcherService is an array, not an
ActiveRecord::Relation.