Add support for deactivating users
Users are authenticate via the DfE Active Directory and authorised
locally.

In the scenario when a user should no longer access the application but
stays within the department, we need a way to disable users and prevent
them accessing the application.

This is also useful to remove users from list in the interface,
preventing them filling up with irrelevent data.