User accounts can be disabled
With the work done to support the disabling of a user account, we want
to allow other users to disable user accounts within the service.

As we want to present a user disabled state as a checkbox, we wrap the
attribute in the database (`distabled_at`) in another accessor
`disabled`.

The disabled method is written to take the paramters from the form (1
and 0 strings) OR true or false.

Setting disabled to a true value, sets disabled_at to the current
time and date.

Setting disabled to a false value, sets disbaled_at to nil.