Users can be deactivated
We do so by wrapping the deactivated_at attribute in an `active` method.

Getting active will return true or false based on the value of
deactivated_at.

Setting `active` with set `deactivated_at`.

Active is the used in the User edit form, essentially giving users the
ability to activate or deactivate users with the date and time of the
change being stored in the database.