Risk protection arrangement csv export service
This takes a slightly differenet approach to exporting a csv file
from the existing class.

We want to handle the scenario when it is passed no projects by
peoducing only the headers, i.e. the instance doesn't care about the
projects passed.

As before we want to only define the headers and order once and have the
object take care of the rest.

The magic comes from using the 'presenter' pattern to provide a way for
a project to be returned in the format the csv requires.

We could split the presenter out at some point soon (perhaps as part of
a refacotor of the FundingAgreementLettersCsvExporterService?)