Create radio button action for type of grant
We need to record the type of sponsored grant the school / trust is
applying for.

This is because the grant team need to know this information, so that
they know which projects they are waiting on grant claim forms for and
can chase delivery officers or case workers.

This data currently sits in KIM, which is being replaced