Handle errors in the establishment fetcher
We saw some instances of this in production, we think this is caused by
the academies api timing out whilst fetching in bulk.

Here we check for any errors in the response and raise if there is one.
We have also added a belt and braces log to Application Insights.