Add paged to trust prefetcher
This is just like the establishment prefetcher work except it updates
the trust fetcher, we'll repeat the commit message:

We use this pre-fetcher to greatly improve speed in the applcation.
The idea being instead of calling the Academies API once for each
project in a view, we can call it once and return multiple results.

We later introduced batching to the object to allow it to prefetch even
greater numbers of records in situations like preparing csv exports and
some of the 'by x' views.

We think when used in combination with the the paging gem (Pagy) the two
collided, resulting it errors that were hard to pinpoint, but likely due
to the offsets used in both objects resulting in empty request to the
API.

For now, we will work around this issue by having a batch prefetcher and
a paged one.

Because the pager only loads at most, 20 records we can comfortable
request those on the API.

When we are not paged and want to get high numbers of records (csv
export, etc) we can still used the batched version.

This is not ideal as we have to remember to choose the correct
behaviour, but right now we need a fix and we should really tackle the
prefetching issue accross the whole codebase as it is pretty much used
everywhere - there is likely a better solution.

We took some time to tidy and dry out this spec here, we tend to mock
the Academies API a new way each time we use it, but I think that is
expected, that said keeping the mocking code close to the test feels
preferable.