Only count in progress projects on the Team projects by user page
On production we noticed that some users on the "By User" Team projects page
were showing a project count, but once we clicked through to see the user's
projects, the number was different. This was because on the user's projects
page, we are only showing in-progress projects; but on the By user page we are
counting *all* assigned projects.

Only count projects which are in progress to bring both pages into alignment.
We aren't interested in counting completed projects.