Fetch projects for export object
The intention here is that in specific context we want to grab a bunch of
projects and prefetch the Academies API entties for them.

When used to fill in a file (or a view that has no pager) we
want to 'batch' the API requests up.

Combining these along with the correct scopes in methods that document
which export they are used for feels like a sound approach.

Nothing is complicated here, we jsut encapsulate the buisiness logic of
what projects to include along with prefetching specifically for the
exports.