Risk Protection Arrangement export UI
We want to offer users a way to get a wide enough view of the data they
are exporting.

Here we setup what we hope to be the pattern for this: an index view
that covers from last month to 4 months in the future, a total of six
months.

The view links to the confirmed and revised project lists, to allow the
user to view the projects that will be included in the export - this
helps build the understanding of how the two are linked, this could be
expecially useful whilst we are migrating to the new product.

There is then a link that triggers the export itself - this already
exists.