Add some delegation to Project
When dealing with projects we often want think about indirect attributes
being on the Project model, e.g. the local authority feels like an
attribute on a project, but is a relation on the establishment (for good
reason!).

Here we add some delegate methods that pass through the messages to the
appropriate objects - this makes dealing with the attributes a little
simpler at the cost of some mis-direction.

We plan to use these delegates very soon, they are not used in this
commit.