Add writer for academy object
An academy on a converison project is exactly the same as the
establishment on a generic project - it simply represents the school
details now it is an academy.

So we can inject the dependancy of academy in later tests, here we add a
wirter method so we can pass in the academy Establishment.