Add academy presenter module
To keep the lines of code to an acceptable level, we are breaking the
various entities that make up the project presenter for csv into
modules.

This is the one that focuses on the academy which is just an 'establishment'.

You might see this as lots of duplication, but the end goal is that the
exports will use these to generate the values for the csv rows so we
would have a log of logic to check if it is the school or academy or
this.