Add director of child services presenter module
To keep the lines of code to an acceptable level, we are breaking the
various entities that make up the project presenter for csv into
modules.

This is the one that focuses on the director of child services, which is
actually from the local_authority on the project.establishment - but via
delegation, we can call it from the project itself.