Add local authority csv presenter  module
To keep the lines of code to an acceptable level, we are breaking the
various entities that make up the project presenter for csv into
modules.

This is the one that focuses on the local authority. The local authority
has be delegated on the Project model, so we can simply reach it without
going through the establishment, even though in reality it must.