Rename caseworker role
This used to represent both a group of users and a permission, now we
have a team attribute, this no longer makes sense and our code is less
approachable.

Here we rename the attribute to what it means:

> This user can be assigned and work on a project

For now, we just want to rename the attribute and have the tests pass,
later we will address some other issues around user behaviour.