Rename regional_delivery_officer role
Like caseworker, this used to represent both a group of users and a
permission, now we have a team attribute, this no longer makes sense and
our code is less approachable.

Here we rename the attribute to what it means:

> This users can add new projects

For now, we just want to rename the attribute and have the tests pass,
later we will address some other issues around user behaviour.