Rename service_support role
Like caseworker, this used to represent both a group of users and a
permission, now we have a team attribute, this no longer makes sense and
our code is less approachable.

Here we rename the attribute to one of the three permissions users in
the group have:

> This user manage user accounts

We have added the other permissions:

- user can manage local authorities
- user can mange conversion urns

For this commit we simply want things to be working, later work will
refine thing further based on these new persmissions.