Change the assigment of permissions
We have renamed all the permissions we have, here we make our model
better reflect reality where both Regional caseworkers and Regional
Delivery Officers work on project i.e. can be assigned to them via the
`assigned_to` relation.

We don't actually enforce this that I can see, any user could be
assigned to a project, but we prvent this via the interface that only
offers 'assignable' user to be selected during assignment.