Migrate assign_to_project data
We have refactor this attribute to mean what it says, which is:

> return true when the user can be assigend to projects

We want all users in the following groups to be 'assignable':

- regional casweworkers
- regional delviery officers

This data migration updates regional delivery officers, who will not
have the attribute set to true.

Unfortunately not all users have a team in production (we are waiting
for them to update this information themselves) but we can still rely on
the `add_new_project` value as only regional delivery officers will have
this set to true right now.

There is no down for the migration as we simple run the service to
update the users.