Add a new Academies API pre fetcher service
We've got in an bit of a mess with all the prefetching we have to do to
speed up the applicaiton interface.

In summary, fetching multiple entities from the Academies API is far
quicker than having to do a single request-response for every project.
We 'inject' the fetched establishments and trusts back into the projects
and then use them in views.

We have quiet a few different ways to do this, and multiple failure
states. This works aims to become the only way to pre fetch Academies
API data.

I does so in batches of 10 and can handle eager loading of associations,
we have had issues with both of these in the past.

Unfortunately, there is no fast way to create enough projects to really
test this other than creating > 10 projects in the database, for me, it
is worth the speed penalty.

Once again, I am mocking the API client in the spec to keep it close to
hand and document what is actually happening - it maybe we can come up
with a way to do this better, for now I would like to keep this.

We won't get rid of the old idividual establishment_fetcher_service and
trust_fetcher_service as those are still in use. Once we are confident
this approach is working well, we plan to swap over.