Use new Acaademies API pre fetcher on team views
Some of these views are currently broken when there are more than 20
projects. Switching to the new pre fetcher service should resolve this.

We've kept the switch to turn off pre fetching on the
ByMonthProjectFetcherService because testing with it on requires a great
deal of mocking, we see the switch as a kind of dependancy inhjection
that allows us to test the soring behaviour without the prefetching.