Add DfE Frontend javascript
This will make the mobile menu work.

I can't see value in a test for this as we would have to launch a whole
browser for this one thing - if you think this is a bad call, please go
ahead and add the spec! We've been running with it broken for weeks now.

We eslint ignore this as it is not our code, we just want to drop in
later versions without worrying about modification.