Update all projects in progress
This now uses the latest Academies API pre fetching.

We also delete the voluntary and sponsored views as we no longer support
those as concepts.

We cleaned up the spec around this - we didn't feel tesing each type of
user could view this was that usefult - we are confident in out policy
specs which should control access to the views.