Update All projects > completed
This now uses the new Academies API pre fetcher.

Removed the voluntary and sponsored views as those are not concepts we
support.

Updated the spec, not sure of the value in testing all the different
users here, we can rely on policy specs for that.

Added a spec for paging and to trigger any Bullet (N+1) warnings.