Update the Academies API pre fetcher
We don't actually want to update the projects with the fetched entities
until we have them all.

It's only the API request that we want to batch up.

Doing this also seems to have cured a lot of Bullet issues we were
seeeing and means we don't have to skip Bullet.