Update All projects > Local authorities
We want this to use the new Academies API pre fetching.

We've moved all the code that gets the right projects into the service
object, however, we have to leave the pre fetching until after the pager
is invoked, this seems to be a rule, we think it is because the pager
returns a new subset of projects that no longer have the pre fetched
entities.