Show the academy order type on the opening list
The academy order type and later the two requires improvements are the
main indicators of a a project that users look for.

Here we add the academy order type as the first step, we are not yet
recording the two requires improvement data.

I've been a little cheeky naming the spec here - I know we will be
changing it from 'Opening' to 'By month' to support both a converison
and a transfer, so I've used the new naming.