Show academy order type on team opening view
Just like the all projects version, the team opening list will soon
become 'by month' and also show the academy order type.

We didn't have to do any work to make this change as we reuse the all
projects view partial, but here is a feature spec to prove it works.

We've also started to hint at how we can organise our feature specs in a
way that mirrors the application.