Project has a significant date
Our Project model should represent the common elements of both
conversion and transfer projects. Both project types have a
'significant' date that works the same way:

- it starts provisional
- it gets confirmed
- it may change
- it is important to users

For a converison project the date is the 'conversion date', for a
transfer project the date is the 'transfer date' - this is the only real
difference.

So that we can model this date on both, we want to raise it up into the
parent Project class from the Conversion::Project class and rename it
along the way.

In this commit we create a new concern, showing that a Project is
concerned with the significant date. We could have rolled this into the
Project class itself, but this way we can encapsulate all the behaviour
in the concern and test it.

With the concern in place, a Project, which can be either a Conversion
or a Transfer now has a significant date, thie paves the way for us to
rename the actual columns in the database and remove the code from
Conversion Project without breaking anything.