Begin to remove conversion date behaviour
This is being replaced with generic significant date behaviour across
both Conversions and Transfers.

For now the column names are still 'converison date' and we have aliased
the `by_conversion_date` scope to keep things happy.

That makes this commit mostly focused on mgirating the other scopes and
ensuring eveything is working.