Rename conversion date histories
Now we have the significant date models, it's time to start changing the
database!

Here we rename the conversion date histories table and related
association on Note.

After this, it's a matter of teaks to being things back in line.

We've left a lot of 'conversion date' lanaguge around becuase it is
actually valid:

> the significant date of a converison project is the conversion date

This will keep the amount of change in this commit low for reviewing, we
can come back and decide how we label things later.