Rename converison date
We've done a bunch of work to make a conversion projects converison date
into a generic 'significant date'.

Here we rename the two columns that store the date and if it is
considered provisional.

We use attribute aliases:

https://api.rubyonrails.org/classes/ActiveModel/AttributeMethods/ClassMethods.html#method-i-alias_attribute

so that most of the existing code still works, this also lets a
Converison::Project have a conversion_date and a Transfer::Project have
a transfer date or you can access generic Projects with
significant_date, whichi is good enough for now.