Add very basic Create form for Transfer Projects
This form collects the absolute minimum for transfers, creates a transfer
project and redirects the user to an empty page on success.

I have not added a show page because in conversions, we do not use the show
page - instead the user is redirected to the tasks page for the project.

Because tasks are out of scope for this ticket, there is no tasks page to
redirect to. Once we have added one, we can change the logic in ProjectsController
to redirect the user to the tasks page for a transfer project.