TasksController works for transfers
We've decided that the complexity for displaying a transfer or a
conversion should be in the view layer. Doing so means we can share the
controller layer between the two project types with only slight changes,
this work starts to make those changes.

This decision does mean duplicaiton of view files, but that keeps the
logic out of the views, which is a good thing.

We lift the routing concern out of conversions namespace and into the main
namepace. This means updating all the path helpers that relate.

We then create a Projectable (I known!) controller concern that does two
things that all controllers concerned with a view of a single projects
neeed:

- load the project, into the @project instance variable
- let Rails know where to find the view templates

The reason for 'Projectable' here is that 'Project' would clash with the
main model - maybe antoher reason for the convention?

The view templates hint will let Rails locate the existing view
templates and later, transfer and conversion specific ones.

With the concern ready, we add it to all the contollers that can use it,
thereby empowering them to deal with transfers as well as conversions.

The final step is to introduce just enough view code to
render a simple transfer project and prove this all works -
that will be next.