Support notes for transfers
We have the approach for all these project sub sections now. Here we
allow a transfer to render it's notes.

The majority of note views are identical across the two project types,
we only have to split out the index views so we can render the
appropriate project summary, the other views stay where they are and
continue to work.