Support internal contacts for Transfers
We have the approach for all these project sub sections now. Here we
allow a transfer to render it's internal contacts.

The majority of these views are identical across the two project types,
we only have to split out the index views so we can render the
appropriate project summary, the other views stay where they are and
continue to work.

We also make the action `index` inline with all the other sub views - in
principle this is the index view of all internal contacts.

A partial keeps this one as DRY as we can.