Significant date is not required
Significant date is declared on a Project and then aliased on
converisons and transfers.

Right now the new transfer form does not collect the transfer date
(significant date) so for now we remove the validation.

Conversions declare conversion date as required as will transfer
eventually - this is only while we build out transfers.