Update service support project views
This commit updates the columns shown on the service support specific
project views;

We remove the route because it is not a concept we want to support any
longer.

We remove school type and phase as  whilst they are useful during the
conversion academy URN journey, they are not needed here.

Overall this user journey is blocked by not having 'unpublished' GIAS
data anyway and until that is resolved, we can do no more.