Add a super basic stakeholder kick off task
We haven't decided what the rest of this task will be yet, but we do
know it will ask users to confirm the transfer date.

Here we copy that behaviour from the almost identical conversion task,
it does mean an almost dupliocate TransferDatable concern and spec, but
we think that is okay for now. We might be able to combine the two into
SignificantDatable at some point.