Update the your project in-progress view
We need this view to support transfers and conversions and we know it
has some issues.

This works aims to strike a middle ground, allowing transfers to be
viewed and improving the view by showing useful infomration with an
improved layout.

We have to ignore a bullet check on one spec as Transfers do not yet
have an all_conditions_met - but we are confident they will.

Regional delivery officers do not work on projects across regions so
that value is not shown to them.