Move all shared create project form attributes
Everything shared belongs in the parent form class.

We have to add the handover note body to the transfer new form to keep
the tests happy - this is not working properly yet, but that will come
in a later commit, right now we want to move the attributes and stay
green.