Change handover notes validation
Previously we required a handover note AKA handover comment.

Strictly speaking this is true, but the way we present the field links
it tightly to the handing over to regional casework services that I
think it is hard to make the case for that validation.

Here we only require the note if the project is being handed over.

In the case where the project stays with the regional team but not the
user, they can add a note manually.

We are still building out the transfer projects views so this is still
not user facing for a transfer.

This does begin to duplicate the form specs for conversion and transfer
- but we will live with that for now.