Update `ByTeamProjectFetcherService` to return Transfer and Conversion projects
All the methods on this service now return Transfer and Conversion projects.
Update the views to use "significant date" instead of "conversion date",
because Transfer projects do not have a conversion date.

Add a count of transfer projects to the `users` method on this service to show
how many transfer projects a user has assigned to them.