Do not send mails to deactivated users
Do not send Team Leader mails or Assigned To mails if the user is not active.

Theoretically you cannot select an inactive user on the Assigned to page, but
this is "belt & braces".

We don't add this restriction to the new user mailer because in theory a new
user shouldn't be deactivated (maybe??)