Move ProjectStatistics into a new Statistics module
We are about to add some User statistics to the service. The current statistics
are all for projects, and it doesn't seem right ot add user statistics to the
same model. So, we will be adding UserStatistics - but first, let's create a
Statistics module to hold all the statistics models.