Data migrate sponsored_support_grant_type
We have been saving the value of this attribute as a string with
hyphens, which is pretty unusual in a Ruby project.

Here we migrate all the values to use underscores, we will then ensure
that only the three accepatble values are allowed, this will strengthen
us against a potential injection attack.