Move `all_conditions_met?` up to Project
Previously, the `all_conditions_met?` value for a Project was determined by its
TasksData, on the `conditions_met_confirm_all_conditions_met` task.

Move this value up to be a native attribute on Project. This reduces a lot of
eager loading and makes the Project model simpler, although it does mean that
the "Conditions met" task becomes another "special case" task.