Bring the flash messages inline with GOV.uk styles
The flash banner should be in its own two-thirds layout.

The banner should be styles as per the GOV.uk guidelines. The main message
must be an h3, and any other text in a paragraph.

We have added a helper to parse the flash messages correctly. If a message only
has one line (no HTML), the helper encapsulates the message in an h3 tag. If
a message contains HTML, we output the HTML in ts entirety. We will have to
ensure any HTML added to flash messages via local files has the correct HTML
format.

We only currently have one multi-line flash message, and the HTML for this has
been updated in the locale file.