(Fix) Enable Transfer projects to be assignd to people & teams
We found a big with assigning Transfer projects to people and teams: the
AssignmentsController was expecting parameters with a `conversion_project` key.
Obviously, with transfer project,s this key will be `transfer_project`.

This solution doesn't feel very elegant but does work. We will probably have to
think of a more elegant solution as we add more project types.