(Fix) Fix alignment of notification banner on Sign in page
The notification banner was mistakenly not put into its own `govuk-grid-row`
div, which meant on some pages the `h1` content was floating next to the
notification banner.

Move the grid row and column divs into the Nootification banner partial, so
that they only appear when a flash is present.

Move the sign-in content into a grid row and column div to preserve the
correct alignment.