Update taam, by user links
We want all master/details views to link on the first column rather than
a specific 'view detail' link.

Here we update the team, by user table. We've also added the
`govuk-link--no-visited-state` which stops the link entering the visited
state (purple) as we don't see the benefit in this context, more info in
the design system here:

https://design-system.service.gov.uk/styles/typography/#links

For this we moved the shared partial as it is not actually shared, it is
specific to the team scope as the team column is not required (it is
implied)