Update by local authority links
We want all master/details views to link on the first column rather than
a specific 'view detail' link.

Here we update the by local authority table. We've also added the
`govuk-link--no-visited-state` which stops the link entering the visited
state (purple) as we don't see the benefit in this context, more info in
the design system here:

https://design-system.service.gov.uk/styles/typography/#links