Update docker-compose file
We HAVE to use docker to be able to run the Microsofy  SQL Server, so it
make sense to run all the 'backing services' (database and Redis) inside
containers and give developers a nice sciprt interface to manage them.

This is how I like to run backing services anyway, we have lots of good
tooling to manage Ruby and Node, Docker makes managing the actual
services like the DB and Redis much easier as the versions can be
controlled per application.