New checks workflow
The idea of this is to run all the checks in CI that mean work can be
merged.

We already have a supporting Dockerfile and docker-compose to provide
the containers.

Along with a clean up, we use the new GitHub Actions cache to speed the
CI process up. We don't know at this point if we'll see the same
improvements with different code as it will effect the caching, that is
why we are adding this new workflow along side the existing ones to
gather some data.

https://docs.docker.com/build/ci/github-actions/cache/#github-cache

The basic idea is to run a job that builds the image and caches it on
GitHub, each job after that builds the image again, but utiliese the
cache, speeding up the process.

Jobs that rely on the compose file build the image again, but this is
also cached.

Essentially, we only need to build the image once, it is then loaded
from the cache each time, but we still have to go through the steps of
building the image to see the benefit.

Initial testing shows ~2 minutes improvement over the exisitng checks.

Eventually we hope to replace the old CI workflows with this single new
one, but we need to validate the cache behaviour first.