Improve the checks Dockerfile
We've moved the installation of firefix and geckodriver into the base
stage but made the installation conditional based on the RAILS_ENV. This
is a common pattern in the Dockerfile.

The result is that we should see the install cached by Docker saving
lots of time in CI.

We can also drop the 'test' stage as all it now does is copies the test
files into the image - we can live with that overhead being in the
production build.

Now the entire build is controlled by the RAILS_ENV build argument.