Validate that the "establishment main contact" for a project is a School contact
We have added a relation between Projects and Contact::Projects called
"establishment main contact". This indicates if a contact is the "main contact"
for an establishment.

We need to make sure that only contacts of category "School" are added as
establishment main contact. To do this, add a validation to make sure that
1) The contact is a "school" contact
2) The establishment main contact id on the project is the same as this
contact id
3) The project is not nil (Contacts do not have to have a project).

This code is a little odd because we are validating against an association, and
the association is not available at `contact.project`, hence having to search
out the Project "up one level" at ::Project (we are currently in another model
called `Project` - `Contact::Project`.

I feel this could be refactored a little.