Allow users to set a School contact as the "establishment main contact"
Now that we have the concept of an "establishment main contact", and a validation
that runs when one is created, we can allow the users to check a checkbox
when creating a contact, to indicate the contact is the "main contact" for the
establishment (school).

The "establishment main contact" must be a "School" category contact.

On the contacts view page, there is an indicator that the contact is the
"establishment main contact".

The establishment main contact status can be added to or removed from an
existing contact.