Tweak checks Docker file
We tried to outpur the coverage report for SonarCloud to consume, but it
just doesn't work.

The paths in the report are that of the container and don't match those
of the job that runs the SonarCloud checks. We tried to `sed` them but
then SonarCloud stopped working and we do not have the capacity to sink
more time.

If you reading this because we are trying again, note a bind mount was
simpler to work with anyway.