Add basic GIAS Establishment model
We want to model GIAS establishment so we can:

- have access to the unopen academy data
- allow searching
- improve performance
- break dependancies on the Academies API

This is a very simple start to that work that creates the table,
indexes, model and factory.

We don't validate any attribute other than `urn` because we do not own
the data (we'll be importing it later) so we need to remain flexible for
the moment and validate the data set.

If a urn was missing, the record would be unusable for us, so we do
validate that attribute.

We try to balannce following the attribute names from the dataset, which
is the GIAS data from a signed in user with all fields selected, whilst
keeping thing approachable.

We store attriubute name and 'code' as these may be useful later, we
already model the 'GOR Region' codes on Project for example.

For the factory, we create a 'with_diocese' trait as one of the main
indicators we use (showing diocese and marking some task as applicable).

Right now all of this is speculative, but takes us on a path we have
designed.